+++ fixtures/tutorial/remix.config.js
@@ -3,5 +3,10 @@ module.exports = {
   browserBuildDirectory: "public/build",
   publicPath: "/build/",
   serverBuildDirectory: "build/app",
-  devServerPort: 8002
+  devServerPort: 8002,
+  routes(defineRoutes) {
+    return defineRoutes(route => {
+      route("*", "catchall.tsx");
+    });
+  }
 };